Optimizations in observerFunctions.m

+++ Parser/outputC/dxdt.h
@@ -35,6 +35,6 @@
 
 int rhs(realtype t, N_Vector y, N_Vector ydot, void *f_data);
 int sensRhs (int Ns, realtype t, N_Vector y, N_Vector ydot, N_Vector *yS, N_Vector *ySdot, void *user_data, N_Vector tmp1, N_Vector tmp2);
-int fJac(long int N, realtype t, N_Vector y, N_Vector fy, DlsMat Jac, void *user_data, N_Vector tmp1, N_Vector tmp2, N_Vector tmp3);
+int fJac(int N, realtype t, N_Vector y, N_Vector fy, DlsMat Jac, void *user_data, N_Vector tmp1, N_Vector tmp2, N_Vector tmp3);
 
 #endif
