+++ Parser/outputC/mexG.c
@@ -78,7 +78,7 @@ void mexFunction (int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]) {
   #endif
   
   #ifdef FJAC
-  int                   (*fj)(long int N, realtype t, N_Vector y, N_Vector fy, DlsMat Jac, void *user_data, N_Vector tmp1, N_Vector tmp2, N_Vector tmp3) = &fJac;
+  int                   (*fj)(int N, realtype t, N_Vector y, N_Vector fy, DlsMat Jac, void *user_data, N_Vector tmp1, N_Vector tmp2, N_Vector tmp3) = &fJac;
   #endif
   
   if (nrhs == 0) {
